coral: enable airbrush prior to shutdown or reboot
This change builds upon ag/6129811 but expands the range
of signals on which we enable airbrush to all 'shutdown' events.

Some of the events that were not caught were:
on property:sys.powerctl=
    "reboot,adb"
    "reboot,bootloader"
    "reboot,recovery"
    "reboot,-p"
    "reboot,"
    and potentially others.

Interestingly, 'init' has a 'shutdown' trigger for all shut-down
related events that I was not aware of.
cs/android/system/core/init/reboot.cpp?q=reboot.cpp&dr&l=588

Bug: 123062257
Change-Id: Ia61f4d4e0cc93eaac05bf2e16b08ad0ec1588803