init: Revert TPU shutdown crash avoidance measures, leave Airbrush off
Remove temporary changes to avoid crashes at shutdown when TPU device is
removed while the Airbrush device is in suspend.  The TPU driver has
been fixed to check PCIe link status and avoid touching the device
during remove if the link is unavailable.

This change reverts the following previous changes:

    72add82 coral: enable airbrush prior to shutdown or reboot
    c4a1e24 coral: enable airbrush prior to shutdown

Bug: 123062257
Test: manual: unbind TPU driver while Airbrush in suspend, reboot
Change-Id: I1e2a758d5a2ba16b1a65299d607cde1810ac1850