vibrator: Apply Calibration to External Control
Use the global scaling knob to apply voltage calibration to externally
controlled vibrations. Additionally, use this for long-vibration to
simplify the setAmplitude() logic. This implementation has the
unfortunate side-effect of also scaling GPIO triggered vibrations that
preempt these vibrations, but in practice the GPIO is only used when the
device is asleep.

Bug: 124701081
Test: Monitor dig_scale/cp_dig_scale during and after vibrations.
Change-Id: Ib04b4e76f8400eac952067d3778e4ea7e1fc1f90
Signed-off-by: Harpreet "Eli" Sangha <eliptus@google.com>