config.fs: init: Add AID_VENDOR_QRTR for qrtr-ns
(1)
Add the AID_VENDOR_QRTR AID in config.fs. Change qrtr-ns to use this
AID for system and group. This AID is reserved for applications that
perform control operations on the QRTR socket family.

The QRTR node ID is hardcoded to 1 within the kernel. Do not configure
the node from userspace. This allows qrtr-ns to drop NET_ADMIN
capabilities.

(2)
change AID_VENDOR_AIRBRUSH from 2906 to 2907 as b/121996920#8

Bug: 121996920
Test: Basic function workble(Dial/SMS/Browser Web/Mobile Data).

Change-Id: I64a9cf0e21e69ad9a7ab63e0981f92232b8a21b5