C2F2: add battery capacities to power_profile.xml
This change adds a single field - battery capacity - to the
power_profile.xml files for C2F2 devices.  This is only the very first
part of populating those files with all of the values that they need to
contain, but it will enable various diagnostic capabilities in tools
like Historian that display power consumption data relative to battery
capacity.

Bug: 129987966
Test: ran on device and verified that data properly propagated into
bugreports.
Change-Id: I162acfae08043405580e203b8d61c0b88e936b20