dumpstate: Fix Maxim FG and WLC path
Fix incorrect WLC path in dumpstate:
/sys/devices/platform/soc/880000.i2c/i2c-1/1-0061/

search on regmap by dumpstate hal requires debugfs directory level
permissions which is frowned upon. Hence provide the actual paths
of regmaps which are needed to be dumped. Query the hw version to
figure out the regmaps that are needed to be dumped.

Reviewed at:
https://googleplex-android-review.git.corp.google.com/c/device/google/coral/+/7388292

Bug: 132044630
Test: adb bugreport
Change-Id: I1e447f4e9fbb870142ca44d2b1daa8846c41f074