usb hal: Switch to DRP after mode changes
It is possible that after active Mode Changes, the port could make a PD
explicit contract. Since the Port Role was changed to DFP or UFP, the
role changes (PR_SWAP or DR_SWAP) will be refused. To avoid this, change
the Port Role back to DRP after active Mode Changes.

Bug: 139163790
Test: build pass; port_type was set to DRP after mode changes
Change-Id: I0d77d7312b09347fcd2c65708cfad7574cb45abf