init.hardware.rc: adjust permissions for oslo tx power calibration
Note that these files are also protected by SELinux such that only the
the OsloFeedback service will be able to read these files.  The
OsloFeedback service is running as a normal Java application so there
are no specific user and groups that can be used to control access to
these files.

Bug: 138146309
Test: flick, reach, and presence on device
Change-Id: I6ce8563c8d188f4a33c62e3b81532ae8ad5ded8a
Signed-off-by: Jeffrey Carlyle <jcarlyle@google.com>