init.hardware.rc: Change per_mgr start timing
The property "persist.vendor.modem.esim.reset" will be set to "needed"
for NFC HAL to reset esim if esim got stuck issue. For this case, we need
to postpone booting modem until NFC HAL finished.
As boot completed, we should force to boot modem if the property cannot
be updated to "done".

Bug: 137376072
Bug: 139319524
Change-Id: I48a44e6ba150eaab3e05ab60a5d861746d854f55