Monitor the property ota.warm_reset
The property is set to inform kernel to do a warm_reset on the next
reboot. This is useful to persist the logs to debug device boot
failures. More details in http://go/rvc-ota-persist-logs.

The property is read by vendor_init. And according to its value,
vendor_init writes a particular sysfs file to schedule a warm reset
on the following reboot.

The vendor init also schedules the warm reset if the current slot hasn't
been marked as successful during boottime, i.e.
ro.boot.slot_successful=no

Bug: 143489994
Test: check the property be read by vendor_init, check the value of the sysfs file
Change-Id: I5f2cde4aed6c21386dbf544700ca4e5119ef75e6