Make corner path coral/flame overlays untranslatable
The overlays for coral and flame corner paths were not marked as
translatable="false", meaning that when the language was set to
cibarA [XB] the paths were reversed. This caused a SystemUI
crashloop at runtime.

Making the path config strings untranslatable fixes the issue.

Bug: 144674035
Bug: 144773362
Test: reproduced crash loop by flashing to master HEAD, setting
language to cibarA [XB], and rebooting the phone. Added parameter
translatable="false" and flashed again; verified that phone works
normally.

Change-Id: I86638593457fa820d8eb3e87a5691d3e33deb0fa