dumpstate: Include Citadel status in all bugreports
Gather Citadel's firmware version, uptime stats, and motherboard type in all
bug reports. None of this can be used to identify a specific phone, so it's
safe to run in production builds.

Previously we were running --id and --selftest. Because these could be used
to indentify a specific Citadel chip, we only gathered it on userdebug and
test builds. We used that to isolate problems with early Citadel samples in
prototype phones, which we've done, fixed, and replaced. We don't need that
info anymore.

Bug: 141235263
Test: manual
Change-Id: I3ea75569a4ec6b0d7c4ef0a3d4b159e06eafc901
Signed-off-by: Bill Richardson <wfrichar@google.com>