Move Telephony vendor overlays to device
The Mainline Telephony overlays were put in the vendor folder.  This worked for
internal builds but not AOSP builds which build out of the device folder.
This commit moves the overlays for Coral and Flame from vendor to device. Internal
builds work with aosp builds so this won't be a problem.

Bug: 148390250
Test: Built coral, aosp_coral, flame
Test: Flashed coral, aosp_coral
Change-Id: Ib88d2e3b01d5e2c9f5316114096642e2d2af8512