Add hints for Adaptive Battery CPU Throttling
Add hints allowing Adaptive Battery to limit the frequency of the big
cores to 50%, 60%, 70%, 80%, or 90%. These will be used for
experimentation first, and only the values that we determine to not
detremental impact on the user experience will potentially be used in
production.

In order to not change how the existing camera and launch hints
behave, the frequency values for the new hints have lower priorities
than the existing ones.

For CPUBigPlusClusterMaxFreq, the new hints use 2015999 instead of
2016000. This is so that they don't use the same values as the
CAMERA_STREAMING_XXX hints. The current cpufreq driver on coral rounds
values up (not down) to the nearest available scaling frequency,
thus allowing the CPU to run to up to 2016Mhz (see b/158134249).

See go/powerbrain-cpu
Bug: 151080800
Test: make and flash, use TurboAdapter and a custom version of Turbo

Change-Id: I88fa86cf9630d64bc71da762c3fcb1299fdc09b1