coral: Reduce rounded corner radius to fix QS clipping path
Now that we have a light QS that could potentially be rendered on top of
a dark app, users can see that the radius of the fancy clipping path is
slightly too high and thus causes visible pixel bleed in the top corners
at mQsExpansion progress = 0. Reduce the radius by 5px (determined
with the help of DEBUG_DISABLE_SCREEN_DECORATIONS) to fix the clipping
issue.

Change-Id: If02ea01fc11247243e4ff8403d8d6bb84b1c0dab